fix: wrong data type for imdbVotes

+++ src/main/kotlin/de/amklee/monomovie/JustWatch.kt
@@ -383,7 +383,7 @@ data class Backdrop(val backdropUrl: String? = null)
 @Serializable
 data class Scoring(
     val imdbScore: Float? = null,
-    val imdbVotes: Int? = null,
+    val imdbVotes: Float? = null,
     val tmdbPopularity: Float? = null,
     val tmdbScore: Float? = null,
     val tomatoMeter: Int? = null,
